extract returnJsonArray out of NodeExecutionFunctions
