+++ packages/core/src/execution-engine/node-execution-context/execute-context.ts
@@ -25,7 +25,6 @@ import {
 
 // eslint-disable-next-line import/no-cycle
 import {
-	returnJsonArray,
 	copyInputItems,
 	normalizeItems,
 	constructExecutionMetaData,
@@ -44,6 +43,7 @@ import {
 import { getDeduplicationHelperFunctions } from './utils/deduplication-helper-functions';
 import { getFileSystemHelperFunctions } from './utils/file-system-helper-functions';
 import { getInputConnectionData } from './utils/get-input-connection-data';
+import { returnJsonArray } from './utils/return-json-array';
 
 export class ExecuteContext extends BaseExecuteContext implements IExecuteFunctions {
 	readonly helpers: IExecuteFunctions['helpers'];
