+++ packages/core/src/execution-engine/node-execution-context/execute-single-context.ts
@@ -14,7 +14,7 @@ import type {
 import { ApplicationError, createDeferredPromise, NodeConnectionType } from 'n8n-workflow';
 
 // eslint-disable-next-line import/no-cycle
-import { getRequestHelperFunctions, returnJsonArray } from '@/node-execute-functions';
+import { getRequestHelperFunctions } from '@/node-execute-functions';
 
 import { BaseExecuteContext } from './base-execute-context';
 import {
@@ -23,6 +23,7 @@ import {
 	getBinaryDataBuffer,
 	getBinaryHelperFunctions,
 } from './utils/binary-helper-functions';
+import { returnJsonArray } from './utils/return-json-array';
 
 export class ExecuteSingleContext extends BaseExecuteContext implements IExecuteSingleFunctions {
 	readonly helpers: IExecuteSingleFunctions['helpers'];
