+++ packages/core/src/execution-engine/node-execution-context/webhook-context.ts
@@ -19,15 +19,12 @@ import type {
 import { ApplicationError, createDeferredPromise } from 'n8n-workflow';
 
 // eslint-disable-next-line import/no-cycle
-import {
-	getNodeWebhookUrl,
-	getRequestHelperFunctions,
-	returnJsonArray,
-} from '@/node-execute-functions';
+import { getNodeWebhookUrl, getRequestHelperFunctions } from '@/node-execute-functions';
 
 import { NodeExecutionContext } from './node-execution-context';
 import { copyBinaryFile, getBinaryHelperFunctions } from './utils/binary-helper-functions';
 import { getInputConnectionData } from './utils/get-input-connection-data';
+import { returnJsonArray } from './utils/return-json-array';
 
 export class WebhookContext extends NodeExecutionContext implements IWebhookFunctions {
 	readonly helpers: IWebhookFunctions['helpers'];
