+++ packages/core/src/node-execute-functions.ts
@@ -107,7 +107,7 @@ axios.interceptors.request.use((config) => {
 	return config;
 });
 
-export const validateUrl = (url?: string): boolean => {
+const validateUrl = (url?: string): boolean => {
 	if (!url) return false;
 
 	try {
@@ -923,30 +923,6 @@ export async function httpRequestWithAuthentication(
 
 		throw new NodeApiError(this.getNode(), error);
 	}
-}
-
-/**
- * Takes generic input data and brings it into the json format n8n uses.
- *
- * @param {(IDataObject | IDataObject[])} jsonData
- */
-export function returnJsonArray(jsonData: IDataObject | IDataObject[]): INodeExecutionData[] {
-	const returnData: INodeExecutionData[] = [];
-
-	if (!Array.isArray(jsonData)) {
-		jsonData = [jsonData];
-	}
-
-	jsonData.forEach((data: IDataObject & { json?: IDataObject }) => {
-		if (data?.json) {
-			// We already have the JSON key so avoid double wrapping
-			returnData.push({ ...data, json: data.json });
-		} else {
-			returnData.push({ json: data });
-		}
-	});
-
-	return returnData;
 }
 
 /**
